+++ routes/user.json
@@ -1,6 +1,6 @@
 
 {
-        "firstName": "Simon ",
+        "firstName": "John",
         "lastname": "Peter",
         "age" : "66"
 }
