+++ routes/users.js
@@ -1,64 +1,17 @@
 import express from "express";
-import { v4 as uuidv4 } from 'uuid';
+import { createUser, getUsers, getUser, deleteUser, updateUser } from '../controllers/users.js';
 
-const router = express.Router(); // Use an Express router 
+const router = express.Router();  
 
-// all routes in here are starting with /users
-router.get('/', (req, res) => {res.send(users);});
 
-let users = []
-// Getting all Users.
-router.get('/', (req, res) => {
-        // console.log ('users');
-
-        res.send(users);
-});
- // adding users into the database.
-router.post('/', (req, res) => {
-         
-        const user = req.body;
-
-        // Generate a unique ID for the user
-
-        users.push({ ...user, id:uuidv4() });
-         
-        res.send(`User with the name ${user.firstName} added to the database`);
-    });
-    
-    // Getting the User ID.
-router.get('/:id', (req , res)=> {
-        const { id } = req.params;
+router.get('/', getUsers );
+ 
+router.post('/', createUser);
+     
+router.get('/:id', getUser );
         
-         // Find the user with the specified ID
-        const foundUser = users.find((user) => user.id === id);
+router.delete('/:id', deleteUser );
 
-        res.send(req.params);
-});
-        // Route to delete a user by ID
-router.delete('/:id', (req, res)=>{
-        const { id }= req.params;
+router.patch('/:id' , updateUser )
 
-        users = users.filter((user) =>user.id !== id);
-
-        res.send(`User with ID ${id} deleted from the database.`);
-});
-// Route to update user in database
-router.patch('/:id' , (req, res)=>{
-        const { id } = req.params;
-
-        const { firstName, lastName, age } = req.body;
-
-        const user = users.find((user) => user.id ===id);
-
-        if(firstName) user.firstName = firstName;
-        
-        if(lastName) user.lastName = lastName;
-        
-        if(age) user.age = age;
-        
-
-        res.send(`User with the ID has been updated`);
-        
-})
-
-export default router;  
+export default router;
